Remove test_reshape.py
